Massive rewrite to add support for UE 5.4. - Several API changes that require modifying dependent code - New extension points added to allow custom save class serialization

+++ UeSaveGame/DataTypes/Text.cs
@@ -53,7 +53,7 @@ namespace UeSaveGame.DataTypes
 			Value?.Deserialize(reader, version);
 		}
 
-		public long Serialize(BinaryWriter writer, PackageVersion version)
+		public int Serialize(BinaryWriter writer, PackageVersion version)
 		{
 			writer.Write((int)Flags);
 			writer.Write((sbyte)HistoryType);
